New parser for Adobe's Flex language.
Flex files are either *.mxml or *.as.

MXML are XML based files with mxtags:
    <mx:Script>
    ...
    </mx:Script>

ActionScript files (*.as) are similar to Javascript
and can also be embedded within MXML files using
the <mx:Script> tags.

The initial file began with the jscript.c parser, 
so a lot of code still needs to be removed and cleaned up.

The following tags are currently supported:
	'f', "function"
	'c', "class"
	'm', "method"
	'p', "property"
	'v', "variable"
	'x', "mxtag"


git-svn-id: https://ctags.svn.sourceforge.net/svnroot/ctags/trunk@682 c5d04d22-be80-434c-894e-aa346cc9e8e8