JavaScript Parser updates for bug 2777310.
var x = 1;
var y = [];
var z = {};

Only x and y are detected - z is ignored,
despite this being perfectly valid syntax
(and quite common, when initialising
a variable for use as a hash table).

[It's equivalent to z = new Object(), in a
similar fashion to y = [] being equivalent
to y = new Array() ].

As far as I can tell, the cause seems to be that
object literals are not being considered as valid
for the right hand side of an assignment.

Now, global variables for z is generated for the release after 5.8.



git-svn-id: https://svn.code.sf.net/p/ctags/code/trunk@763 c5d04d22-be80-434c-894e-aa346cc9e8e8