Parsing javascript quotes like this:       "('adding '+a+' and ' +b);\    // separate string using "\"
Starts with a " and has a " at the end of the line, though that is part of a
comment and I assume the \ after the ; means it is a line continuation
indicator and the rest of the line should be ignored.  So, I have updated
parseString() to do just that.

This correctly parses:
    Test\jsFunc_tutorial.js

git-svn-id: http://svn.code.sf.net/p/ctags/code/trunk@798 c5d04d22-be80-434c-894e-aa346cc9e8e8
(cherry picked from commit 7b84378496fba3e09d0b1d439a11952d1ffd4e06)
Signed-off-by: Masatake YAMATO <yamato@redhat.com>