Introduce ctags.conf.d
With this patch ctags load all files named *.ctags and *.conf under
/etc/ctags.conf.d, /usr/local/etc/ctags.conf.d and ~/.ctags.d as configuration
files. This feature is useful to manage user own language supports.

Consider the case you are a member of a programmer team, and all
members of the team are using ctags. When you define a language support for
Swine, something peculiar language but the main language of the project.

Without the new feature, you have to say "please, copy and paste this
language support to your .ctags". The user must use an editor or cat command to
introduce a new language support written in --langdef, --langmap, and --regex-*
options.

With the feature, you can say "please, put this language support to
/etc/ctags.conf.d/swine.ctags." The user can use mv or cp command to introduce
it. Updating the support is also easy; just replacing the file.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>