modeline: choose a parser by emacs mode specification at the first line
If a parser for a given file cannot be determined by file extension,
a editor mode line can be used as a good hint for the choice.

emacs detects following two formats at the first line of a file as a
mode specification:

     -*- MODE -*-[1]

     -*- mode: MODE; ... -*-[2]

[1] https://www.gnu.org/software/emacs/manual/html_node/emacs/Choosing-Modes.html#Choosing-Modes
[2] https://www.gnu.org/software/emacs/manual/html_node/emacs/Specifying-File-Variables.html
This patch makes ctags determines a parser by this mode specification.

TODO: getPatternLanguage is not considered yet.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>