refactor: unify functions for checking with nameLanguage and that with extensionLanguage when choosing a language parser
getExtensionLanguage was prior to getNamedLanguage. However, we should
use both impartially.

Now ctags can detect shbang line, emacs mode and vim filetype. In many
cases the name of emacs mode or vim filetype is same as the
name of parser.

e.g.

If a file has -*- python -*- line as the head, codes in
getEmacsModeLanguageAtFirstLine will pick up "python" and
choose a language parser getExtensionLanguage.

Though "python" is not registered as an extension.

    extern parserDefinition *PythonParser (void)
    {
	static const char *const extensions[] = { "py", "pyx", "pxd", "pxi" ,"scons",
						  "python2", "python3",
						  NULL };

"python" is the name of language parser.

	parserDefinition *def = parserNew ("Python");

Newly introduced getAnyLanguage checks both
LanguageTable [i]->currentExtensions and  LanguageTable[i]->name
to find a language parser.

Adding the name of parser to the extensions is another approach.
However, we have to change wider area of codes in the approach.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>