refactor: introduce a function calling both nominateLanguageCandidates and getTwoGramLanguage
Currently language parsers candidates are not considered in
getEmacsModeLanguageAtFirstLine, getInterpreterLanguage,
getEmacsModeLanguageAtEOF and getVimFileTypeLanguage.  We should do
the same in those functions as in file extension handling.

As the first step I make the code for handling file extension generic
enough for applying above functions.

In this patch getSpecLanguage is introduced. getSpecLanguage wraps
sequencial invocations of nominateLanguageCandidates and
getTwoGramLanguage.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>