store the key for searching TgTableEntry to candidates array
In nominateLanguageCandidates, only LangType was stored to candidates
array of parsers. However, LangType is not enough informative to find
TgTableEntry because a language parser can have multiple TgTableEntry
objects. A TgTableEntry can be find with a key. Here key is assumed as
a file extension, parser name or something string.

To pick proper TgTableEntry for the nominated language parser, not
only LangType but also the key must be stored to candidates array.
This patch introduced parserCandidate strcut type for extending the
candidates array to store the keys for each candidate.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>