php: free memory blocks allocated in a function as function local var
A memory block is allocated in initPhpEntry and is assigned to
fullScope the function local variable. This is not any kind of memory
leak. However, for simplify the output of valgrind, in this patch, the
variable is moved to the out of the function as a file local vairable.
The memory block is alocated at php's initialize method and freed at
its finalize method.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>