introduce config file search path
ctags can load configration files which are specified with --options
option. A path for an existing file or directory was expected as
the argument for --options.

In this patch config file search path are introduced.  For argment of
--options, before searching existing fie or directory resolved
relatively from current working directpry, config file search path is
searched.

Currently config file search path can be specified with
CTAGS_DATA_PATH enviroment variable. Paths in CTAGS_DATA_PATH are
searched to load a configuration file. Like PATH enviroment variable,
the paths can be enumerated with `:' as a separator.

Let me describe following command line:

    $ cd /home/devel
    $ CTAGS_DATA_PATH=/a:/b ./ctags --options=x ...

To load x as configuration file (or directpry), ctags looks at /a/x
then /b/x. If x cannot be found, finally ctags looks at /home/devel/x.
If you want to load /home/deve/x anyway, use one of following command line:

    $ CTAGS_DATA_PATH=/a:/b ./ctags --options=./x ...

or

    $ CTAGS_DATA_PATH=/a:/b ./ctags --options=/home/devel/x ...

If a directory is givens as an argment for --options, files which
names are ended with ".ctags" or ".cnf" are loaded alphabetically
order. Loading are done in recursively; directories which names are
ended with ".d" are loaded, too.

TODO: update ctags.1.
Signed-off-by: Masatake YAMATO <yamato@redhat.com>