regex: introduce a flag to cut off the rest of matching operations
The regex language parser tries matching all expressions with ALL lines of
given files regardless one of expression is actually matched or not.

This patch introduces a flag `!'. If an expression with ! is matched
with a line, the regex language parser doesn't try the rest of
expressions with the line.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>