consider .in as a suffix for a template file
Like Makefile.in and foo.c.in sometimes .in is used as a suffix for a
template file.

This patch adds a new rule choosing a language parser; ".in" part at
the end of input file name is striped before comparing the extension
of file name and before matching the pattern of file name.

e.g.

	consdier input file name is "Makeflie.in".
	There is no language parser for the file name.
	With this patch Make language parser can be chosen
	from the striped file name "Makefile".

	consdier input file name is "foo.c.in".
	There is no language parser for the file name.
	With this patch C language parser can be chosen
	from the striped file name "foo.c".

The new rule is added to the end of rules in getFileLanguage.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>