unify simplify lang parser choosing mechanism
When choosing a lang parser, a file name is tested twice: once with
file extension and once with file name pattern.  This patch integrates
choosing lang parser with file extension chooser into that with file
name pattern.

As the result the code is simplified; and conflictions with the two
tests can be arbitrated with single getTwoGramLanguage call.

Now objc-tg-corpus.d test case runs succssfully.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>