introduce alias mechanism for language name
This patch introduces --<LANG>-alias= and --list-aliases=
options.

An alias is alternative name for a lang parser.  When choosing a lang
parser refers the contents of input file to extract an interpreter
name, emacs major mode information or vim surtax name. The alias is
matched with the extracted string to choose a parser.

The extracted name was compared with language own file extension
tables and language parser names. The language parser names are
suitable for this purpose. However, file extensions are not suitable.

e.g. Consider "#!/usr/bin/python3" is found in an input file named
input "input.nolang".

You may want ctags to choose python language parser from the "python3"
automatically. Before introducing this alias mechanism, you have to
add "python3" to extension list like:

    static const char *const extensions[] = { "py", "pyx", "pxd", "pxi" ,"scons",
                                             "python3", NULL };

However, python3 is not an extensions. Even if you add python3 to
extension list, there is no serious bad effect. However, for making
ctags applicable to more complex situation, two concepts "extension"
and "the name of interpreter" are better to be separated.

An alias for a LANG can be defined with aliases field of
parserDefinition struct or --<LANG>-alias= option. The alias is
treated as a glob pattern when matching a string extracted from input
file.

e.g.

	--python-alias="python[23]"

Signed-off-by: Masatake YAMATO <yamato@redhat.com>