add --<LANG>-map= option as an alias for --langmap= option
(This is nothing to do with alias mechanism introduced by
 --<LANG>-alias option.)

There are three formats for defining or customizing a language parser:

      A. --langmap=<LANG>:..., and

      B. --regex-<LANG>=..., and

      C. --<LANG>-foo=...

If these formats are unified, writing configuration library (*.ctags
file) becomes a bit easier.

This patch unifies A into C; --<LANG>-map= option is added as an alias
for --langmap=.  --langmap= is not removed for keeping compatibility.

With the alias

     --langmap=mib:(*-MIB*)

can be written like:

    --mib-map=(*-MIB*)

Signed-off-by: Masatake YAMATO <yamato@redhat.com>