add --<LANG>-regex= option as an alias for --regex-<LANG>= option
(This is nothing to do with alias mechanism introduced by
 --<LANG>-alias option.)

There are three formats for defining or customizing a language parser:

      A. --langmap=<LANG>:..., and

      B. --regex-<LANG>=..., and

      C. --<LANG>-foo=...

If these formats are unified, writing configuration library (*.ctags
file) becomes a bit easier.

This patch unifies B into C; --<LANG>-regex= option is added as an
alias for --regex-<LANG>=.  --regex-<LANG>= is not removed for keeping
compatibility.

With the alias

     --regex-mib=/^([^ \t]+)[ \t]+DEFINITIONS ::= BEGIN/\1/d,definitions/!

can be written like:

    --mib-regex=/^([^ \t]+)[ \t]+DEFINITIONS ::= BEGIN/\1/d,definitions/!

Signed-off-by: Masatake YAMATO <yamato@redhat.com>