allow disabling a kind defined in regex attached to built-in language parser
ctags allows adding extra regex to built-in language parser.

Folowing is taken from scripts/tags.sh of linux kernel:

	 --regex-c='/^SYSCALL_DEFINE[[:digit:]]?\(([^,)]*).*/sys_\1/' \

Further more a kind can be defined to each regex.

Issue is that a kind defined in regex attached to built-in language
parser could not be disabled with --<LANG>-kinds=- option. This patch
allows it.

NOTE: A kind defined in regex attached to language defined with
--langdef can be disable.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>