refactor: remove unnessary references to regex field of parserDefinition
I'm working on adding --<LANG>-cmd= option to ctags.  This will be
alternative way to define new language parser to --<LANG>-regex=.

Some of ctags code expects --<LANG>-regex as the only way to
extend ctags. This will be barrier for adding --<LANG>-cmd
feature.

Referencing regex field of parserDefinition is used in such code. I
reviewed the usage of regex field. In this patch I remove unnessary
references to regex field of parserDefinition.

This patch removes the references from resetLanguageKinds and
printKinds.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>