support regex long flag own parameter
In

	987ce1b introduce long flag of --regex-<LANG> option

I introduce long flag. This patch extends the feature.
The flag either long or short can represent just a boolean value, on
or off. With this patch, a long flag can take one string argument.

   --regex-<LANG>=.../[flag1=PARAMETER1][flag2=PARAMETER2]...

If a long flag is ended with `=', flag evaluator of ctags considers
the long flag takes a parameter. String after `=' till `]' is
refered as a parameter and passed to the handler of long flag.

TODO: character `]' cannot be used in parameter.
Signed-off-by: Masatake YAMATO <yamato@redhat.com>