introduce readLineWithNoSeek
readLine is useful but cannot be used for pipe because the function
expects a given stream is seek-able; fgetpos and fsetpos are used internally.

readLineWithNoSeek does the same as readLine do but doesn't
expect a given stream is seek-able. readLineWithNoSeek is
useful for reading lines from pipe which is connected to
an external command.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>