check supported features in unit tests
A unit case can be skipped with this patch.

If a test case requires a .feature of ctags and the target ctags
doesn't support the feature, the test case is automatically skipped.

Features required by a unit test case must be listed in
Units/TEST/features. Features supported in a ctags file can
be listed with --list-features option.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>