introduce better patch fixing command 'endlessloop' of .vim input from fedora
This patch deals the same issue fixed by

7fb36a2f4690374526e9e7ef4f1e24800b6914ec (from exuberant)
and/or
e59325a576e38bc63b91abb05a5a22d2cef25ab7 (from Debian).

This patch is in ctags package of Fedora.
The file name of patch is ctags-5.7-segment-fault.patch.

This patch is introduced because this is better than
above two patches in two aspects

1. this patch doesn't tag command 'endlessloop'.
   String surrounded by single quotes after command
   is not valid syntax as a vim input.

   A patch from debian tags the string.

2. this patch doesn't conflict with
   2fe9fd54564fedb89314ded58e496e687d8a7513.

   A patch from exuberant conflicts with it.