remove SYS_INTERPRETER ifdef condition
SYS_INTERPRETER is used as build-time condition whether #!/bin/sh
style notion at the head of a script should be supported or not.

However, even if the platform where ctags runs is not supported the
notion, a platform whether script files given as input to ctags can
supported the notion. In any platform ctags should be able to
handle more various type of input.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>