check file existence when expanding a directory name on search path list
When --options=foo is specified, foo.d is checked whether
it is a directory or not. If it is a directory, files under
the directory are loaded as config files.

For checking isDirectory function is used but is had a bug.
It was only isDirectory filed of return value of eStat.
exists was not checked. If exists is FALSE, the value
of isDirectory field is not meaningless.

With this patch isDirectory function checks exists field
before checking isDirectory field.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>