prepare DESTDIR variable to switch the instillation directory
This patch is taken from ctags package of Fedora.

With this patch you can switch the top directory where ctags and
related files are installed. This is helpful to make a package like
rpm in Fedora. You can use the variable with following command line:

	$ make DESTDIR=$buildroot install

NOTE: DESTDIR is independent from --prefix= of configure script.

The concept of DESTDIR is also used in automake. See info of
automake for more details.