prepare the way to gather test cases for known bugs
We expect tests under Units are passed. If you see "FAILED",
you think there is something wrong.

With the patch we can gather test cases for known bugs.
Gathered test cases are all failed but we know they
waill be failed because we have not fixed bugs associated
with the tests yet. In this case the result message "FAILED"
is not what we want.

Use ".b" as suffix instead of ".d", the message is changed
to "failed but KNOWN bug" even if the test case is failed.

NOTE: Don't add too complext and many test cases as ".b".
Remember ctags is not any kind of compiler or interpreter.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>