introduce preload directories
This feature is similar to ~/.ctags but it is implemented
by utilizing directories.

Instead of editing a file, you can customize the bahavior of ctags
with puting a file to one of preload directory or deleteing a file
from one of preload directory. This feature is very package system of
GNU/Linux friendly.

On GNU/Linux, following directories are used for preloading.

    ~/.ctags.d
    /etc/ctags/preload
    /usr/share/ctags/preload

.ctags files and .d directories under above directories are loaded
when ctags starts.

The implementation is based on a patch submitted http://sourceforge.net/p/ctags/bugs/341/.
This implementation is inspred by systemd.

TODO: automatic test for this feature
Signed-off-by: Masatake YAMATO <yamato@redhat.com>