suppress the output from read shell-builtin command
In testing.mak read shell built-in command is used.  On bash if a
parameter for read is omitted, REPLY variable is used for the place
where read data is stored. This behavior is different from dash.

read of dash prints something. (I have not inspected the output.)
To avoid the output, give a variable name to read shell built-in
explicitly in this patch.

dash is used widely Debian and its variants.
Even on Fedora you can test this patch with SHELL=/bin/dash.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>