sql: fix memory leaks related to setjmp/longjmp
readToken frequently called in sql parser uses longjmp.  The control
jumps to the top level of sql parser.  As the result memory objects
allocated callers of readToken can be lost references.

With using trash, all memory objects are tracked.  They are freed in
trashMakeEmpty at the top level. So all deleteToken and vsringDelete
for memory objects traced by the trash are removed.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>