Units: fix a bug in matching the lines of language file and ctags --list-languages output
If a unit test expects a language parser and target ctags executable
doesn't support the language, the unit test should be skipped.

languages file under each unit test case represents the language
the test case expects ctags to have. However, matching code
that compares languages file and the output of --list-languages
was broken. As the result a unit test case is reported as failed
though it should be skipped.

This patch removes a space in output of --list-languages.
Let's consider "coffee" language.
If a coffeetags command is not available, the parser should
be disabled.

  $ ./ctags --list-languages
  coffee [disabled]

In testing.mak, this output is used in a for loop.
The for loop picked up the output of "ctags --list-languages" word by
word. So "coffee" and "[disabled]" are picked up separately. This is
the reason why matching didn't work expectedly. With using sed command
the for loop picked up the output of "ctags --list-languages" line by
line("coffee[disabled]").

Reported by @b4n in #60.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>