units: suppress the error report of shell itself
When ctags is crashed, the shell process running the
ctags process reports the crash:

    Testing verilog-sv-nocontext misc/units: line 203: \
    14080 Segmentation fault ${_CMDLINE} 2> "${ostderr}" > "${orawout}" \
    failed (unexpected exit status: 139)

This report disturbs the well-formatted output of tests.

With this patch, ctags is run in a subshell; and the subshell the
stderr and stdout are redirected to SHSTDERR.tmp and SHSTDOUT.tmp.  As
the result the crash report doesn't disturbs the well-formatted output
of tests but can be refereed as a separated file later if needed.

Issue reporeted by @vhda in #60.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>