xcmd: extend xcmd protocol to deal with the case back-end command is unavailable
This patch fixes a bug reported by @kaushalmodi in #131.
coffee xcmd driver dealt the error of coffeetags, its back-end command.
But it assumed coffeetags is always available. As the result it
tries to launch coffeetags command even if the command is not available.

In this change, coffee xcmd driver checks the available of back-end; and
report unavailability with special status code 77.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>