Revert "javascript: don't add property if it is a regexp."
A regex is not a valid JavaScript property name, so remove specific
code to handle it.  Although this was added to prevent trying and
generate an empty tag, the parser has plenty of paths leading to this,
all of which are a lot more common than invalid syntax.

Anyway, a fix for generating empty tags should be more generic and not
handle one single invalid case.

This reverts commit c906e24872308eb8d5a8a5f648685b6b469da396.

Also, fix the test cases not to use this invalid syntax for no good
reason, as they were only supposed to check for regex handling.