Escape control control characters when writing tag fields
Nothing prevents the various tag fields (from name to signature through
file name) from containing arbitrary data, including control characters
like `\t` or `\n`.

Some parsers, like the JavaScript one, have limited handling of such
input to mitigate possible problems, but this is often not enough (it
for example won't help with file names) and is fragile as each parser
has to do it manually.

I see two possible approaches:

1. Filter the offending bytes when creating the tag entry, in
   `makeTagEntry()`.  I see two implications worth noting with this
    approach:
    1. It requires string duplication that may be avoided with other
       solutions.
    2. It is an arbitrary limitation of the in-memory format, which may
       or may not be wanted (e.g. when making CTags a library).
2. Filter the offending bytes when writing the tag entry.  This has the
   advantage of only affecting the output format (the very goal of the
   change) and not the internal representation, so it has very low
   implication on how internals handle these.

This patch is a mostly working attempt at the second approach.  It
isn't completely ready yet because it doesn't handle `%Ns` and the like
that are use in 2 places (see the `FIXME`), but is a proof of concept.

The implementation here might *not* be the one we want to go with
because it basically re-implements `fprintf()`, which is a non-trivial
function for non-trivial cases: e.g. support for `%-Ns` is not trivial.

The advantage is that it requires minimal changes on how the output is
generated, and that the escaping being implicit lowers a lot the risk
of forgetting to escape some parts.  It also avoids having to duplicate
strings to build the escaped version, as filtering can be done straight
when producing the actual output in most cases.