Add a new --undef command-line option
This option allows to disable the generation of macro tags from #undef
directives, see PR #221 for the rationale. The default behaviour is
unchanged, i.e. a tag is still generated.

The help and man have been updated accordingly, and I added a small unit
to check that --undef=no disables the generation of a tag. The
directives.c unit already checks that a tag is generated when no option
is passed (which is equivalent to --undef=yes).