Auto-detect missing coffeetags under rbenv
Due to the way executables are installed by rbenv, `type coffeetags`
returns true if coffeetags executable is missing in the current ruby
version but is available in one of the other installed rubies.

This commit fixes this issue by explicitly checking rbenv for the
availability of the coffeetags executable within the current ruby
version.