Regenerate the build system whenever its sources changed
Regenerate the build system out of its sources when it changed, to keep
it automatically up to date.  This allows to simply run `make` after
configure.ac has been altered without having to worry about manually
running `autoreconf`.  This is particularly useful when configure.ac
has been updated by a VCS pull, as one is likely not to have noticed
the change.