Accept empty name pattern in regex if exclusive is specified
If an empty name pattern(``//``) is found in ``--regex-<LANG>`` option
ctags warns it as wrong usage of the option. However, the flags
``exclusive`` or ``x`` is specified, the warning is suppressed. This
is imperfect approach for ignoring text insides comments but it may
be better than nothing.

E.g.

	--m4-regex=/#.*(define|undefine|s?include)\>//x
	--m4-regex=/\<dnl.*(define|undefine|s?include)\>//x

Signed-off-by: Masatake YAMATO <yamato@redhat.com>