Unify kinds behind regex patterns
Though a same kind letters are used, multiple
patterns created each own kinds internally.

  $ ctags --langdef=Z --regex-Z=/a/\1/t,test/ \
                      --regex-Z=/b/\1/t,test/ \
		      --list-kinds=Z
  t  test
  t  test

Two kinds which have the same name are listed with
--list-kinds. This is unwanted behavior. The two
kinds should be unifined into one.

As a side effect of this fix, defining a language
becomes a bit simple. Once a kind and its description
is defined, you don't have to write the description
for a pattern for the same kind repeatedly.

You can write patterns for Z like:

  $ ctags --langdef=Z --regex-Z=/a/\1/t,test/ \
                      --regex-Z=/b/\1/t/ \
		      --list-kinds=Z

Signed-off-by: Masatake YAMATO <yamato@redhat.com>