go: parse struct members
Parse struct members such as

struct
{
	a, b, c int
	e []int `annotation`
	int
	*int
}

Special attention has to be paid to anonymous fields (like the last two
in the above struct) which don't have variable name assigned to them
and for which no tag should be generated.

Apart from that the newly added parseStructMembers() function resembles
the existing parseConstTypeVar() (but merging the two would make things
too messy).