csharp, java,vera: fix known bugs recorded under fuzz-hitting-assertions.r
Assertions in csharpTagKind, and families(*TagKind) are too strict using from
includeTag.

includeTag is a predicate; it tells whether the current parser deals
with a given type or not.  It means includeTag receives a type which
the current parser doesn't deal with. When includeTag receiving
non-dealing type, the assertions in *TagKind are false.

The assertions are meaningless for includeTag.

In this commit assertions are disabled when *TagKind
is called from includeTag.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>